+++ buster.js
@@ -1,0 +1,7 @@
+module.exports.unit = {
+  environment: 'node',
+  tests: [
+    'test/cli.js',
+    'test/jenkins.js'
+  ]
+};