+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.0.11 (SNAPSHOT)
-*
+* Move commands setup to conf/commands.json, cli handling to bag.cli.command
+* Move request handling to bag.http.request
+* Change unit tests from Mocha to Buster
 
 ### 0.0.10
 * Add stop command